Realm DB 비동기 처리 개선 (#21)
* Add firebase fabric App ID...

* Update Gradle version...

* Update Gradle version...

* Add Fabric Handler..

* Fix Handler..

* Fix load memo async process..

* Add firebase configuration..

* Fix workflows..

* Fix syntax tab..

* Refactoring code...

* Set callback for realm async transaction...

* Independent async, sync algorithm...

* fix image transaction async methods..

* Refactoring unused import and so on...

* Fix wrong toast message...

+++ app/src/main/java/xyz/neonkid/simplememoj/main/util/RealmHelper.java
@@ -5,6 +5,8 @@ import androidx.annotation.Nullable;
 import java.io.File;
 import java.util.Calendar;
 import java.util.TimeZone;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
 
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
@@ -65,9 +67,10 @@ public class RealmHelper {
      * @param color 메모 색깔
      * @param mi 메모에 들어갈 이미지 (Nullable)
      */
-    public void insertMemo(int id, String title, String content, String color, @Nullable MemoImage mi) {
+    public void insertMemo(int id, String title, String content, String color, @Nullable MemoImage mi,
+                           Realm.Transaction.OnSuccess onSuccess, Realm.Transaction.OnError onError) {
         // 기존 메모인 경우..
-        if (editMemoWithId(id, title, content, color, mi))
+        if (editMemoWithId(id, title, content, color, mi, onSuccess, onError))
             return;
 
         // 새로운 메모인 경우
@@ -81,67 +84,123 @@ public class RealmHelper {
 
         realm.executeTransaction(realm -> {
             if (mi != null)
-                memo.setTransactionImg(mi);
+                memo.setTransactionImg(realm, mi);
             realm.copyToRealm(memo);
         });
     }
 
     /**
-     * 메모에 이미지를 DB에 저장하는 함수입니다.
+     * 메모에 이미지를 DB에 저장하는 동기 함수입니다.
      * 이미지에 ID가 부여 되지 않은 경우, 새 ID가 부여됩니다.
      *
      * 단, Realm DB에서 기존에 부여된 ID가 있다면 Exception을 발생합니다.
      *
+     * @param realm Realm instance
      * @param memo 대상 메모
      * @param mi 추가할 이미지
      */
-    public void insertMemoImage(Memo memo, MemoImage mi) {
+    public void insertMemoImage(Realm realm, Memo memo, MemoImage mi) {
         if (mi.getId() == 0)
-            mi.setId(getNewMemoImageId());
+            mi.setId(getNewMemoImageId(realm));
 
         final MemoImage managedMi = realm.copyToRealmOrUpdate(mi);
         memo.getImgs().add(managedMi);
     }
 
     /**
-     * DB에 저장된 모든 메모를 가져오는 함수입니다.
+     * DB에 저장된 모든 메모를 가져오는 비동기 함수입니다.
      * 작성 순서별로 배치하기 위해 Primary Key룰 부여받은 순으로 정렬합니다.
      *
      * @return 저장된 메모들
      */
     public RealmResults<Memo> getMemos() {
+        return realm.where(Memo.class).findAllAsync().sort(Memo.FIELD_ID, Sort.ASCENDING);
+    }
+
+    /**
+     * DB에 저장된 모든 메모를 가져오는 함수입니다.
+     * 작성 순서별로 배치하기 위해 Primary Key룰 부여받은 순으로 정렬하며,
+     *
+     * 반드시 해당 클래스에서 동기 처리를 위해서만 사용해주십시오
+     *
+     * @param realm Realm Instance
+     * @return 저장된 메모들
+     */
+    private RealmResults<Memo> getMemos(Realm realm) {
         return realm.where(Memo.class).findAll().sort(Memo.FIELD_ID, Sort.ASCENDING);
     }
 
     /**
-     * 한 메모에 들어있는 모든 이미지를 가져오는 함수입니다.
+     * 한 메모에 들어있는 모든 이미지를 가져오는 비동기 함수입니다.
      *
      * @param memoId 대상 메모 ID
      * @return memoId에 해당하는 이미지들
      */
     public RealmResults<MemoImage> getMemoImages(int memoId) {
+        return realm.where(MemoImage.class).equalTo("memo.id", memoId).findAllAsync()
+                .sort(MemoImage.FIELD_ID, Sort.ASCENDING);
+    }
+
+    /**
+     * 한 메모에 들어있는지 모든 이미지를 가져오는 함수입니다.
+     *
+     * 반드시 해당 클래스에서 동기 처리를 위해서만 사용해주십시오
+     *
+     * @param realm Realm Instance
+     * @param memoId 대상 메모 ID
+     * @return memoId에 해당하는 이미지들
+     */
+    private RealmResults<MemoImage> getMemoImages(Realm realm, int memoId) {
         return realm.where(MemoImage.class).equalTo("memo.id", memoId).findAll()
                 .sort(MemoImage.FIELD_ID, Sort.ASCENDING);
     }
 
     /**
-     * 메모의 ID 값에 해당하는 메모를 가져오는 함수입니다.
+     * 메모의 ID 값에 해당하는 메모를 가져오는 비동기 함수입니다.
      *
      * @param id 대상 메모 ID
      * @return id에 해당하는 모든 메모
      */
     public RealmResults<Memo> getMemoWithId(int id) {
+        return realm.where(Memo.class).equalTo(Memo.FIELD_ID, id).findAllAsync()
+                .sort(Memo.FIELD_ID, Sort.ASCENDING);
+    }
+
+    /**
+     * 메모의 ID 값에 해당하는 메모를 가져오는 동기 함수입니다.
+     *
+     * 반드시 해당 클래스에서 동기 처리를 위해서만 사용해주십시오.
+     *
+     * @param realm Realm Instance
+     * @param id 대상 메모 ID
+     * @return id에 해당하는 모든 메모
+     */
+    private RealmResults<Memo> getMemoWithId(Realm realm, int id) {
         return realm.where(Memo.class).equalTo(Memo.FIELD_ID, id).findAll()
                 .sort(Memo.FIELD_ID, Sort.ASCENDING);
     }
 
     /**
-     * 메모 이미지의 ID 값에 해당하는 이미지를 가져오는 함수입니다.
+     * 메모 이미지의 ID 값에 해당하는 이미지를 가져오는 비동기 함수입니다.
      *
      * @param id 대상 메모 이미지 ID
      * @return id에 해당하는 모든 메모 이미지
      */
     public RealmResults<MemoImage> getMemoImageWithId(int id) {
+        return realm.where(MemoImage.class).equalTo(MemoImage.FIELD_ID, id).findAllAsync()
+                .sort(MemoImage.FIELD_ID, Sort.ASCENDING);
+    }
+
+    /**
+     * 메모 이미지의 ID 값에 해당하는 이미지를 가져오는 동기 함수입니다.
+     *
+     * 반드시 해당 클래스에서 동기 처리를 위해서만 사용해주십시오.
+     *
+     * @param realm Realm Instance
+     * @param id 대상 메모 이미지 ID
+     * @return id에 해당하는 모든 메모 이미지
+     */
+    private RealmResults<MemoImage> getMemoImageWithId(Realm realm, int id) {
         return realm.where(MemoImage.class).equalTo(MemoImage.FIELD_ID, id).findAll()
                 .sort(MemoImage.FIELD_ID, Sort.ASCENDING);
     }
@@ -158,72 +217,82 @@ public class RealmHelper {
      * @return 저장 결과
      */
     private boolean editMemoWithId(int id, String title, String content, String color,
-                                   @Nullable MemoImage mi) {
-        RealmResults<Memo> res = getMemoWithId(id);
+                                   @Nullable MemoImage mi, Realm.Transaction.OnSuccess success, Realm.Transaction.OnError error) {
+        AtomicReference<RealmResults<Memo>> res = new AtomicReference<>(getMemoWithId(id));
 
-        if (res.isEmpty())
+        if (res.get().isEmpty())
             return false;
 
-        Memo memo = res.get(0);
+        realm.executeTransactionAsync(realm -> {
+           res.set(getMemoWithId(realm, id));
 
-        if (memo != null) {
-            realm.executeTransaction(realm -> {
-                memo.setTitle(title);
-                memo.setContent(content);
-                memo.setColor(color);
-                memo.setModified(Calendar.getInstance(TimeZone.getDefault()).getTime());
+           if (res.get().isValid()) {
+               Memo memo = res.get().get(0);
 
-                if (mi != null)
-                    memo.setTransactionImg(mi);
-            });
-        }
+               if (memo != null) {
+                   memo.setTitle(title);
+                   memo.setContent(content);
+                   memo.setColor(color);
+                   memo.setModified(Calendar.getInstance(TimeZone.getDefault()).getTime());
+
+                   if (mi != null)
+                       memo.setTransactionImg(realm, mi);
+               }
+           }
+        }, success, error);
 
         return true;
     }
 
     /**
-     * 대상 메모를 제거하는 함수입니다.
+     * 대상 메모를 제거하는 비동기 함수입니다.
      * 메모 이미지가 존재한다면, 이미지 제거 후, 삭제를 진행합니다.
      *
      * @param id 대상 메모 ID
      */
-    public void deleteMemoWithId(int id) {
-        RealmResults<Memo> res = getMemoWithId(id);
+    public void deleteMemoWithId(int id,
+                                 Realm.Transaction.OnSuccess success, Realm.Transaction.OnError error) {
         deleteMemoImageWithMemoId(id);
 
-        if (res.isEmpty())
-            return;
+        realm.executeTransactionAsync(realm -> {
+            RealmResults<Memo> res = getMemoWithId(realm, id);
 
-        realm.executeTransaction(realm -> res.deleteAllFromRealm());
+            if (res.isValid())
+                res.deleteAllFromRealm();
+        }, success, error);
     }
 
     /**
-     * 대상 메모 이미지를 제거하는 함수입니다.
+     * 대상 메모 이미지를 제거하는 비동기 함수입니다.
      * 실제 이미지와 동시에 제거하게 됩니다.
      *
      * @param imageId 대상 메모 이미지 ID
      */
-    public void deleteMemoImageWithMemoImageId(int imageId) {
-        RealmResults<MemoImage> res = getMemoImageWithId(imageId);
+    public void deleteMemoImageWithMemoImageId(int imageId,
+                                               Realm.Transaction.OnSuccess success,
+                                               Realm.Transaction.OnError error) {
+        realm.executeTransactionAsync(realm -> {
+            RealmResults<MemoImage> res = getMemoImageWithId(realm, imageId);
 
-        if (res.isEmpty())
-            return;
+            if (res.isEmpty())
+                return;
 
-        for (MemoImage img : res)
-            FileUtils.deleteFile(new File(img.getUri()));
+            if (res.isValid()) {
+                for (MemoImage img : res)
+                    FileUtils.deleteFile(new File(img.getUri()));
 
-        realm.executeTransaction(realm -> {
-            res.deleteAllFromRealm();
-        });
+                res.deleteAllFromRealm();
+            }
+        }, success, error);
     }
 
     private void deleteMemoImageWithMemoId(int memoId) {
-        RealmResults<MemoImage> res = getMemoImages(memoId);
+        realm.executeTransactionAsync(realm -> {
+            RealmResults<MemoImage> res = getMemoImages(realm, memoId);
 
-        if (res.isEmpty())
-            return;
-
-        realm.executeTransaction(realm -> res.deleteAllFromRealm());
+            if (res.isValid())
+                res.deleteAllFromRealm();
+        });
     }
 
     /**
@@ -236,11 +305,25 @@ public class RealmHelper {
         return getAutoIncrementId(new Memo());
     }
 
-    private int getNewMemoImageId() {
-        return getAutoIncrementId(new MemoImage());
+    private int getNewMemoImageId(Realm realm) {
+        return getAutoIncrementId(realm, new MemoImage());
     }
 
     private int getAutoIncrementId(Object obj) {
+        Number curId = null;
+
+        if (obj instanceof Memo)
+            curId = realm.where(Memo.class).max(Memo.FIELD_ID);
+        else if (obj instanceof MemoImage)
+            curId = realm.where(MemoImage.class).max(MemoImage.FIELD_ID);
+
+        if (curId == null)
+            return 1;
+        else
+            return curId.intValue() + 1;
+    }
+
+    private int getAutoIncrementId(Realm realm, Object obj) {
         Number curId = null;
 
         if (obj instanceof Memo)
